do not purge output if making benchmarks

+++ regtest.py
@@ -1185,8 +1185,11 @@ def test_suite(argv):
 
                 match_count += 1
 
-                if suite.purge_output == 1 and not pfile == output_file:
+                if suite.purge_output == 1 and not pfile == output_file and not suite.make_benchmarks:
 
+                    print("pfile:", pfile)
+                    print("output_file:", output_file)
+                    
                     # delete the plt/chk file
                     try:
                         shutil.rmtree(pfile)
